Updated discord-wallet-connection to use discord-library 1.3.0
