+++ discord-wallet-connection/pom.xml
@@ -1,18 +1,18 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
-		<groupId>de.zorro909.blank</groupId>
+		<groupId>com.blank.humanity</groupId>
 		<artifactId>blankdiscordbot-parent</artifactId>
-		<version>1.1.0</version>
+		<version>1.3.0</version>
 	</parent>
 	<version>0.0.1-SNAPSHOT</version>
 	<artifactId>discord-wallet-connection</artifactId>
 
 	<dependencies>
 		<dependency>
-			<groupId>de.zorro909.blank</groupId>
+			<groupId>com.blank.humanity</groupId>
 			<artifactId>discord-library</artifactId>
-			<version>1.1.0</version>
+			<version>1.3.0</version>
 		</dependency>
 		<dependency>
 			<groupId>org.web3j</groupId>
