+++ discord-wallet-connection/src/main/java/com/blank/humanity/discordbot/wallet/entities/DiscordWallet.java
@@ -1,4 +1,4 @@
-package de.zorro909.blank.BlankDiscordBot.wallet.entities;
+package com.blank.humanity.discordbot.wallet.entities;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,7 +6,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
-import de.zorro909.blank.BlankDiscordBot.entities.user.BlankUser;
+
+import com.blank.humanity.discordbot.entities.user.BlankUser;
+
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
@@ -14,7 +16,7 @@ import lombok.NonNull;
 @Entity
 @Data
 @NoArgsConstructor
-public class DiscordWalletSalt {
+public class DiscordWallet {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
@@ -26,6 +28,6 @@ public class DiscordWalletSalt {
     private BlankUser user;
 
     @NonNull
-    private String salt;
+    private String walletAddress;
 
 }
