+++ discord-wallet-connection/src/main/java/com/blank/humanity/discordbot/wallet/entities/DiscordWalletSalt.java
@@ -1,4 +1,4 @@
-package de.zorro909.blank.BlankDiscordBot.wallet.entities;
+package com.blank.humanity.discordbot.wallet.entities;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -7,7 +7,8 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
 
-import de.zorro909.blank.BlankDiscordBot.entities.user.BlankUser;
+import com.blank.humanity.discordbot.entities.user.BlankUser;
+
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
@@ -15,7 +16,7 @@ import lombok.NonNull;
 @Entity
 @Data
 @NoArgsConstructor
-public class DiscordWallet {
+public class DiscordWalletSalt {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
@@ -27,6 +28,6 @@ public class DiscordWallet {
     private BlankUser user;
 
     @NonNull
-    private String walletAddress;
+    private String salt;
 
 }
