+++ discord-wallet-connection/src/main/java/com/blank/humanity/discordbot/wallet/messages/WalletFormatDataKey.java
@@ -1,6 +1,7 @@
-package de.zorro909.blank.BlankDiscordBot.wallet.messages;
+package com.blank.humanity.discordbot.wallet.messages;
 
-import de.zorro909.blank.BlankDiscordBot.utils.FormatDataKey;
+import com.blank.humanity.discordbot.utils.FormatDataKey;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NonNull;
