+++ discord-wallet-connection/src/main/java/com/blank/humanity/discordbot/wallet/rest/DiscordWalletRegistrationDto.java
@@ -1,4 +1,4 @@
-package de.zorro909.blank.BlankDiscordBot.wallet.rest;
+package com.blank.humanity.discordbot.wallet.rest;
 
 import lombok.Builder;
 import lombok.Data;
